design : 바뀐 디자인으로 footer 리팩토링
