+++ src/app/(main)/grouping/page.tsx
@@ -3,9 +3,9 @@ import { getGroups } from '@/apis/groups/apis';
 import { Keys } from '@/apis/groups/keys';
 import { FloatAddButton } from '@/components/Button';
 import { RejectedFallback } from '@/components/common/ErrorBoundary';
-import { BottomNavigationBar } from '@/components/common/NavigationBar';
 import { HydrationProvider } from '@/components/common/Provider';
 import { Spacing } from '@/components/common/Spacing';
+import { Footer } from '@/components/Footer';
 import { Header } from '@/components/Header';
 import { QueryAsyncBoundary } from '@suspensive/react-query';
 import Image from 'next/image';
@@ -25,12 +25,13 @@ export default function Grouping() {
       <QueryAsyncBoundary rejectedFallback={RejectedFallback} pendingFallback={null}>
         <HydrationProvider queryKey={Keys.getGroups()} queryFn={() => getGroups(0)} isInfiniteQuery>
           <GroupingCardList />
+          <GroupingCardList />
         </HydrationProvider>
       </QueryAsyncBoundary>
       <Link href="/grouping/create" className="fixed inset-x-0 bottom-100 mx-auto flex max-w-400">
         <FloatAddButton className="ml-auto" />
       </Link>
-      <BottomNavigationBar page="grouping" />
+      <Footer page="grouping" />
     </>
   );
 }
