+++ src/app/(main)/meeting/page.tsx
@@ -1,5 +1,5 @@
 import ContentSection from './components/ContentSection.client';
-import { BottomNavigationBar } from '@/components/common/NavigationBar';
+import { Footer } from '@/components/Footer';
 import { Header } from '@/components/Header';
 import { redirect } from 'next/navigation';
 import React from 'react';
@@ -24,7 +24,7 @@ export default function MeetingPage({ searchParams }: MeetingPageProps) {
         </Header.Left>
       </Header>
       <ContentSection />
-      <BottomNavigationBar page="meeting" />
+      <Footer page="meeting" />
     </div>
   );
 }
