--- src/components/common/NavigationBar/index.ts
@@ -1,2 +1,0 @@
-export { default as TopNavigationBar } from './TopNavigationBar';
-export { default as BottomNavigationBar } from './BottomNavigationBar';
